+++ packages/account-react-native/src/auth-portal/hooks/use-auth-portal-controller.ts
@@ -23,7 +23,7 @@ export const useAuthPortalController = () => {
   const { connectingAccount } = useSophonContext();
   const { requiresAuthorization } = useSophonContext();
 
-  const shouldAuthorize = isConnected || connectingAccount;
+  const shouldAuthorize = isConnected || !!connectingAccount;
 
   const currentStep = useMemo<AuthPortalStep | null | undefined>(() => {
     switch (method) {