+++ packages/account-partner/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sophon-labs/account-partner
 
+## 3.0.0-next.3
+
+### Patch Changes
+
+- More enhancements on ui and auth flow
+- Updated dependencies
+  - @sophon-labs/account-core@3.0.0-next.3
+
 ## 3.0.0-next.2
 
 ### Patch Changes